core: Fix websocket closins way too fast
It seems it was not sufficient to set the timeout
on the spell caster to be higher... The problem
is a little more _low level_ it seems to come up
at the NAT level which does some sort of "garbage
collection" itself when a connection is not sufficiently
used...

To address this I introduce PING-PONG heartbeats.
I tested this code and have been able to send messages
even minutes later without needing to reload the app.